BDOG-2570 UserManagementConnector
