+++ app/views/OutOfDateTeamDependenciesPage.scala.html
@@ -15,7 +15,7 @@
  *@
 
 @import uk.gov.hmrc.cataloguefrontend.DateHelper._
-@import uk.gov.hmrc.cataloguefrontend.connector.UserManagementConnector.{UMPError, TeamDetails}
+@import uk.gov.hmrc.cataloguefrontend.connector.UserManagementPortalConnector.{UMPError, TeamDetails}
 @import uk.gov.hmrc.cataloguefrontend.ViewMessages
 @import uk.gov.hmrc.cataloguefrontend.connector.model.{Dependencies, Dependency, TeamName}
 @import uk.gov.hmrc.cataloguefrontend.connector.RepoType
