+++ app/views/teams/TeamInfoPage.scala.html
@@ -17,7 +17,7 @@
 @import _root_.views.html.partials.DependenciesTeamPartial
 @import uk.gov.hmrc.cataloguefrontend.DateHelper._
 @import uk.gov.hmrc.cataloguefrontend.{DisplayableTeamMember, ViewMessages}
-@import uk.gov.hmrc.cataloguefrontend.connector.UserManagementConnector.{UMPError, TeamDetails, SlackInfo}
+@import uk.gov.hmrc.cataloguefrontend.connector.UserManagementPortalConnector.{UMPError, TeamDetails, SlackInfo}
 @import uk.gov.hmrc.cataloguefrontend.connector.RepoType
 @import uk.gov.hmrc.cataloguefrontend.connector.model.{Dependency, TeamName}
 @import uk.gov.hmrc.cataloguefrontend.model.Environment
